pyproject.toml
[project]
name = "reserchium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.2",
    "flagembedding>=1.3.5",
    "llama-index>=0.14.15",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-anthropic>=0.10.10",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-postprocessor-flag-embedding-reranker>=0.4.1",
    "llama-index-retrievers-bm25>=0.6.5",
    "llama-index-tools-brave-search>=0.4.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-parse>=0.6.54",
    "networkx>=3.6.1",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "pyvis>=0.3.2",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
[project.scripts]
reserchium = "reserchium.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
